Add:セレクトボックスにドラマを表示
セレクトボックスにドラマを取得し、選択できるようにしました。
